Small day 2 fix

+++ Day2/tip-calculator.py
@@ -3,7 +3,7 @@ Day 2/100 Tip Calculator
 '''
 
 print("Welcome to the tip calculator!")
-bill = int(input("What was the total bill?: $"))
+bill = float(input("What was the total bill?: $"))
 tip = int(input("How much tip would you like to give?: ")) / 100
 people = int(input("How many people to split the bill?: "))
 pay = (bill * (1+tip)) / people
